requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
amqp==2.6.1
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==4.4.7
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
cycler==0.11.0
Deprecated==1.2.13
et-xmlfile==1.1.0
fastapi==0.70.0
flower==0.9.7
fonttools==4.28.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
humanize==3.13.1
hyperframe==6.0.1
idna==3.3
iso8601==0.1.16
jwt==1.3.1
kiwisolver==1.3.2
kombu==4.6.11
matplotlib==3.5.0
networkx==2.6.3
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
passlib==1.7.4
Pillow==8.4.0
priority==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.23
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
pypika-tortoise==0.1.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
redis==4.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
tortoise-orm==0.17.8
typing-extensions==4.0.0
uvicorn==0.15.0
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
